[FIX] buttons CRUD integrate with backend

+++ src/components/contents/buttons.tsx
@@ -5,10 +5,8 @@ import {
   Modal,
   ModalBody,
   ModalContent,
-  ModalHeader,
   useDisclosure,
 } from "@nextui-org/react";
-import { qna } from "@prisma/client";
 import { Trash, Check, X, Pencil } from "lucide-react";
 import { FaqEditModal } from "../modals/faq-edit-modal";
 import { useEffect, useState } from "react";
@@ -16,10 +14,16 @@ import { TextEditModal } from "../modals/text-edit-modal";
 
 export const DeleteButton = ({
   id,
+  setStatus,
   type,
+  session,
+  blockId,
 }: {
   id: number;
+  setStatus: Function;
   type: string;
+  session: any;
+  blockId: number;
 }) => {
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
   const [isLoading, setIsLoading] = useState<boolean>(false);
@@ -27,26 +31,49 @@ export const DeleteButton = ({
   const handleDelete = async () => {
     setIsLoading(true); // Set loading state to true
 
-    try {
-      const res = await fetch(`/api/${type}/${id}`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: "include",
-        body: JSON.stringify({
-          id: id,
-        }),
-      });
+    const role = session.role
 
-      if (!res.ok) {
-        const errorResponse = await res.text();
-        console.error("API Response Error:", errorResponse);
-        throw new Error(`Network response was not ok: ${res.status} ${res.statusText}`);
+    try {
+      if (role === "admin") {
+        const res = await fetch(`/api/${type}/${id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            "Cookie": `session=${session}`,
+          },
+          credentials: "include",
+          body: JSON.stringify({
+            id: id,
+          }),
+        });
+
+        if (!res.ok) {
+          const errorResponse = await res.text();
+          console.error("API Response Error:", errorResponse);
+          throw new Error(
+            `Network response was not ok: ${res.status} ${res.statusText}`,
+          );
+        }
+      } else {
+        const res2 = await fetch(`/api/content/${blockId}`, {
+          method: "PATCH",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          credentials: "include",
+          body: JSON.stringify({
+            id: id,
+            status: "deletePending",
+          }),
+        });
+
+        if (!res2.ok) {
+          throw new Error("Network response was not ok");
+        }
       }
 
-      const result = await res.json();
-      if (result) window.location.reload();
+      window.location.reload();
+      
     } catch (error) {
       console.error(error);
     } finally {
@@ -70,10 +97,18 @@ export const DeleteButton = ({
             <ModalBody className="flex flex-col items-center justify-center py-10">
               <p>Are you sure want to delete this?</p>
               <div className="flex flex-row w-full items-center justify-center gap-5 mt-4">
-                <Button className="w-md" onClick={onOpenChange} disabled={isLoading}>
+                <Button
+                  className="w-md"
+                  onClick={onOpenChange}
+                  disabled={isLoading}
+                >
                   Cancel
                 </Button>
-                <Button className="w-md bg-red-300" onClick={handleDelete} disabled={isLoading}>
+                <Button
+                  className="w-md bg-red-300"
+                  onClick={handleDelete}
+                  disabled={isLoading}
+                >
                   {isLoading ? "Deleting..." : "Delete"}
                 </Button>
               </div>
@@ -90,11 +125,13 @@ export const EditButton = ({
   setStatus,
   type,
   session,
+  blockId,
 }: {
   id: number;
   setStatus: Function;
   type: string;
   session: any;
+  blockId: number;
 }) => {
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
   const [modal, setModal] = useState<JSX.Element | null>(null);
@@ -105,7 +142,14 @@ export const EditButton = ({
         setModal(<FaqEditModal id={id} />);
         break;
       case "text":
-        setModal(<TextEditModal openChange={onOpenChange} session={session} id={id}/>);
+        setModal(
+          <TextEditModal
+            openChange={onOpenChange}
+            session={session}
+            id={id}
+            blockId={blockId}
+          />,
+        );
         break;
       default:
         setModal(null);
@@ -130,10 +174,22 @@ export const EditButton = ({
   );
 };
 
-export function createButton(id: number, setStatus: Function, api: string) {
+export const ConfirmButton = ({
+  id,
+  setStatus,
+  type,
+  session,
+  blockId,
+}: {
+  id: number;
+  setStatus: Function;
+  type: string;
+  session: any;
+  blockId: number;
+}) => {
   const handleUpdate = async () => {
     console.log("create");
-    const res = await fetch(api, {
+    const res = await fetch(`/api/content/${blockId}`, {
       method: "PATCH",
       headers: {
         "Content-Type": "application/json",
@@ -163,9 +219,43 @@ export function createButton(id: number, setStatus: Function, api: string) {
   );
 }
 
-export function cancelButton(id: number) {
+export const  CancelButton = ({
+  id,
+  setStatus,
+  type,
+  session,
+  blockId,
+}: {
+  id: number;
+  setStatus: Function;
+  type: string;
+  session: any;
+  blockId: number;
+}) => {
+  const handleCancel = async () => {
+    console.log("cancel");
+    const res = await fetch(`/api/${type}/${id}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+        "Cookie": `session=${session}`,
+      },
+      credentials: "include",
+      body: JSON.stringify({
+        id: id,
+      }),
+    });
+
+    if (!res.ok) {
+      throw new Error("Network response was not ok");
+    }
+
+    const result = await res.json();
+    if (result) window.location.reload();
+  };
+
   return (
-    <button>
+    <button onClick={handleCancel}>
       <X
         className="mt-1 hover:bg-gray-100 rounded-full p-1"
         width={30}
@@ -175,10 +265,22 @@ export function cancelButton(id: number) {
   );
 }
 
-export function updateButton(id: number, setStatus: Function) {
-  const handleUpdate = async () => {
-    console.log("update");
-    const res = await fetch(`${process.env.NEXT_PUBLIC_URL}/api/faq`, {
+export const CancelDeleteButton = ({
+  id,
+  setStatus,
+  type,
+  session,
+  blockId,
+}: {
+  id: number;
+  setStatus: Function;
+  type: string;
+  session: any;
+  blockId: number;
+}) => {
+  const handleCancel = async () => {
+    console.log("cancel");
+    const res = await fetch(`/api/content/${blockId}`, {
       method: "PATCH",
       headers: {
         "Content-Type": "application/json",
@@ -194,14 +296,14 @@ export function updateButton(id: number, setStatus: Function) {
       throw new Error("Network response was not ok");
     }
 
-    const result: qna = await res.json();
+    const result = await res.json();
     if (result && result.status) setStatus(result.status);
   };
 
   return (
-    <button onClick={handleUpdate}>
-      <Check
-        className="mt-1 hover:bg-yellow-100 rounded-full p-1"
+    <button onClick={handleCancel}>
+      <X
+        className="mt-1 hover:bg-gray-100 rounded-full p-1"
         width={30}
         height={30}
       />
